Types: split single index.ts file into files based on function/category
