+++ packages/types/src/calls.ts
@@ -1,0 +1,11 @@
+export type CallUser = CallUserSlack | CallUserExternal;
+
+export interface CallUserSlack {
+  slack_id: string;
+}
+
+export interface CallUserExternal {
+  external_id: string;
+  display_name: string;
+  avatar_url: string;
+}
