+++ packages/types/src/message-attachments.ts
@@ -1,0 +1,83 @@
+import { PlainTextElement } from './block-kit/composition-objects';
+import { Block, KnownBlock } from './block-kit/blocks';
+
+/**
+ * Add {@link https://api.slack.com/messaging/composing/layouts#attachments secondary attachments} to your messages in Slack.
+ * Message attachments are considered a legacy part of messaging functionality. They are not deprecated per se, but they may change in the future, in ways that reduce their visibility or utility. We recommend moving to {@see Block} constructs instead. Read more about {@link https://api.slack.com/messaging/composing/layouts#when-to-use-attachments when to use message attachments}.
+ */
+export interface MessageAttachment {
+  blocks?: (KnownBlock | Block)[];
+  fallback?: string; // either this or text must be defined
+  color?: 'good' | 'warning' | 'danger' | string;
+  pretext?: string;
+  author_name?: string;
+  author_link?: string; // author_name must be present
+  author_icon?: string; // author_name must be present
+  author_subname?: string;
+  title?: string;
+  title_link?: string; // title must be present
+  text?: string; // either this or fallback must be defined
+  fields?: {
+    title: string;
+    value: string;
+    short?: boolean;
+  }[];
+  image_url?: string;
+  thumb_url?: string;
+  footer?: string;
+  footer_icon?: string; // footer must be present
+  ts?: string;
+  actions?: AttachmentAction[];
+  callback_id?: string;
+  mrkdwn_in?: ('pretext' | 'text' | 'fields')[];
+  app_unfurl_url?: string;
+  is_app_unfurl?: boolean;
+  app_id?: string;
+  bot_id?: string;
+  preview?: MessageAttachmentPreview; // https://api.slack.com/methods/chat.unfurl#markdown
+}
+
+// https://api.slack.com/methods/chat.unfurl#markdown
+export interface MessageAttachmentPreview {
+  type?: string;
+  can_remove?: boolean;
+  title?: PlainTextElement;
+  subtitle?: PlainTextElement;
+  iconUrl?: string;
+}
+
+export interface AttachmentAction {
+  id?: string;
+  confirm?: Confirmation;
+  data_source?: 'static' | 'channels' | 'conversations' | 'users' | 'external';
+  min_query_length?: number;
+  name?: string;
+  options?: OptionField[];
+  option_groups?: {
+    text: string
+    options: OptionField[];
+  }[];
+  selected_options?: OptionField[];
+  style?: 'default' | 'primary' | 'danger';
+  text: string;
+  type: 'button' | 'select';
+  value?: string;
+  url?: string;
+}
+
+export interface OptionField {
+  description?: string;
+  text: string;
+  value: string;
+}
+
+export interface Confirmation {
+  dismiss_text?: string;
+  ok_text?: string;
+  text: string;
+  title?: string;
+}
+
+export interface LinkUnfurls {
+  [linkUrl: string]: MessageAttachment;
+}
